Reflect Go int, uint, and uint32 as format:int64
Go int and uint are 64-bit on every platform the module supports, and
uint32's range exceeds int32 — clients generated from the previous
mapping mis-typed these fields. int8/16/32 and uint8/16 keep int32;
uint64 stays int64 with a documented range caveat (OpenAPI has no
unsigned format).

Spec-affecting: documents that previously said int32 for int-typed
fields now say int64.

Closes #23

+++ internal/schema/schema_test.go
@@ -47,15 +47,15 @@ func TestReflectSchema_IntegerSizes(t *testing.T) {
 		value  any
 		format string
 	}{
-		{int(0), "int32"},
+		{int(0), "int64"},
 		{int8(0), "int32"},
 		{int16(0), "int32"},
 		{int32(0), "int32"},
 		{int64(0), "int64"},
-		{uint(0), "int32"},
+		{uint(0), "int64"},
 		{uint8(0), "int32"},
 		{uint16(0), "int32"},
-		{uint32(0), "int32"},
+		{uint32(0), "int64"},
 		{uint64(0), "int64"},
 	}
 	for _, c := range cases {
